Refactor code - moved some organisms to pages
