+++ src/app/app.module.ts
@@ -16,9 +16,9 @@ import { ReactiveFormsModule } from "@angular/forms";
 import { ToastModule } from "primeng/toast";
 import { ConfirmationService, MessageService, PrimeNGConfig } from "primeng/api";
 import { EditGameComponent } from "./components/pages/edit-game/edit-game.component";
-import { EditSetComponent } from "./components/organisms/edit-set/edit-set.component";
-import { GameSettingsComponent } from "./components/organisms/game-settings/game-settings.component";
-import { PrintComponent } from "./components/organisms/print/print.component";
+import { EditSetComponent } from "./components/pages/edit-set/edit-set.component";
+import { GameSettingsComponent } from "./components/pages/game-settings/game-settings.component";
+import { PrintComponent } from "./components/pages/print/print.component";
 import { SidebarComponent } from "./components/organisms/sidebar/sidebar.component";
 import { SidebarItemComponent } from "./components/molecules/sidebar-item/sidebar-item.component";
 import { ConfirmDialogModule } from "primeng/confirmdialog";
