+++ src/app/components/molecules/kwartet-card-preview/kwartet-card-preview.component.html
@@ -1,0 +1,1 @@
+<div class="cover"></div>
