+++ src/app/components/molecules/kwartet-card-preview/kwartet-card-preview.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-kwartet-card-preview',
+  standalone: true,
+  imports: [],
+  templateUrl: './kwartet-card-preview.component.html',
+  styleUrl: './kwartet-card-preview.component.scss'
+})
+export class KwartetCardPreviewComponent {
+
+}
