+++ src/app/components/pages/game-settings/game-settings.component.ts
